Introduce `site` to split data #6753
The `site` is guess from the server hostname and most data (but not all)
are affected to one single site. A user is also site dependent so a human
will need two logins/password to access both DILPS and Tiresias with
different session.

However a site is not a security feature and data filtering is done by client
only. The server does not enforce the separation. This is to allow to see data
across sites in some cases.