Restore datation search #6848
`type="Email"` is a bit of a trick because `int` would not work. This is
because graphql-doctrine does not support to override the GraphQL type
EqualOperatorInt which is already declared in the lib for scalar.

By declaring it for a "fake" scalar we avoid this limitation, even though
it is not stricly correct.