Restore period search by range #7001
This restore TypeNumericRangeComponent but this time with
always required `from` and `to` values to always define a complete
range. Then we can properly match this user-defined period with existing
periods in DB (which is semantically different that applying 2 independent
date filters).

This basically restore the period searching as it existed in
7183edb37fd4722af0ee472ebcf597acf9e7c7ed