Hide collection whose parent are not accessible #7135
Previously a child collection that had a visibility to `member` was
accessible, but if its parent was `private` then it was not possible
to get the child hierarchicName, because we were correctly not allowed
to access the parent. This was an incorrect attempt to show an incomplete
branch of the collection tree.

Now we a collection is accessible if all its ancestors are also accessibles.
So we can show entire and complete hierarchies, but we cannot show incoherent
partial hierarchies anymore. This also most likely match what end-user want
to do when moving a parent collection from member to private to hide all the
content of the collection.