Drastic reduction of SQL queries during export #7677
Because OneToOne relations forces Doctrine to always fetch both object
on each of the relation, it generate a lot of SQL queries. When fetching
a Card, its potential Change was always fetched. But very often it was
completely useless. This was especially true for export where fetch
thousands of cards but never need changes. So there is no need to generate
thousands of unnecessary queries.

The data in DB is kept consistent and coherent via unique constraint. So
it has no downsides except a small special handling in the Card class itself.
And it has huge benefit for performance. An export of 116'436 cards used to
take 7.41 minutes and 383 MiB. It now takes only 5.7 minutes and 199 MiB.