Site specific config #7901
The entire PHP config can be overridden by site. The
overriding config, if necessary, can be declared in

- `config/autoload/dilps.local.php`
- `config/autoload/tiresias.local.php`

If those files exist they will automatically be
loaded when appropriate.

Because Dilps is the default, and Tiresias is the special
case, the recommended way is to configure Dilps entirely
in `local.php`. And then override only a few value in `tiresias.local.php`.
This way, there is no `dilps.local.php` file at all.

The final, complete configuration will be cached separately
for each site, respectively in:

- `config-cache.dilps.php`
- `config-cache.tiresias.php`

The build process was adapted to delete both, but if those
cached files were being handled by another (manual?) process,
then it will have to change too.