Code can never be updated via suggestion system #7952
Because the field is unique in DB, we cannot reasonably
suggest a value, because the suggested value would always
already exists in DB when we try to accept the suggestion.
So we totally exclude the field from suggestion system.

This also make sense semantically, because code is pretty much
an ID for humans, and we would almost never change an ID.