Enable batch queries and NetworkInterceptor #7984
By using the Angular specific http-batch package we leverage
Angular HttpClient to query graphql, instead of native fetch.
This allow to integrate nicely with HttpClient interceptor mechanism.

The only exception left is when we upload a file. In that case a different
link is used which does not use HttpClient. So only in that case we
"manually" count the queries