Declare all services as readonly
Because pretty much anything that is injected should
never been changed afterwards.

We did see those mis-usage in the past in our code and
that broke typings, because a model service was replaced
by another model service that returned a different shape
of objects. And the compiler was unable to tell us that our
template used something that may or may not exists.