Silently fallback on 0 pagination if asking for negative pagination #8253
Negative offset and limit throw error in MariaDB, because they make no sense at all.
Instead of letting the MariaDB buble up to the human, we silently fallback to an empty
pagination, avoiding errors, but returning nothing at all. This makes the API a little
more forgiving for cases where the pagination comes from the URL and a human forge and
incorrect URL, or when our code incorrectly generate an invalid URL.