Migrate from Protractor and Puppeteer to Playwright #8336
Protractor is EOL, and Playwright can replace entirely Puppeteer (for
our usage for Karma). This simplifies a bit the maintenance since we
can install system deps automatically, and we are sure that everything
is always synced, and because it stores browser binaries outside of
node_modules.

For now, we only test on Chromium, because that's what we always did, but
we could also test other browsers later.