Don't export card twice or more #8569
By splitting into two queries we can avoid the trouble of having
a LIMIT on a JOINed query. This gives us a certainty of non duplicated
cards.

Also move to a cursor based pagination style. While this does not directly
relate to the reported bug, it allows to keep the export consistent if a
card is being deleted while the export is being done.