Only update artist cache of impacted cards #9224
Because DB is huge, it's important to only treat records that may have
been touched, and not the whole DB.

The most complex case is artist deletion where we need to make a list of
touched cards before we actually delete the artist. But the cache update
can only be done after the artist is deleted. To do that we use a
temporary table, `updated_card`, to pass data across triggers.