Debounce statistics recording #9221
Statistics are slightly delayed, with a debounce, to prevent counting a
user that would very quickly click on many collections. This kind of
behavior should not be counted as a real page views. And in a more
general way, it slows the whole user experience by delaying more
important queries to be sent. Now the most useful queries have the
priority.