Fix selecting thesaurus with enter key adds the search term too
When we add a thesaurus from the autocomplete thesaurus component by
pressing the enter key, it adds the search term too.

This behaviour occurs because the "optionSelected" event from the
autocomplete component is triggered not only by a click on the element
in the list but by the enter key too.

We already have an event listener on the input field for the enter
keypress event. It results by both events being catched: one by the
autocomplete component and one by the input component.

The first event catched is the one from the autocomplete event, it will
add the element to the list of selected thesaurus. Then the enter
keypress event is catched and will add the content of the input in the
selected thesaurus list.

The check "this.autocomplete.activeOption" to avoid the key event to add
the content of the input didn't works as the optionSelected event was
already complete and so "this.autocomplete.activeOption" always return
null.

As both of these events are on different components,
event.stopPropagation will not works (they're not the "same" event).

Another way was to disable adding an element from keypress on the
autocomplement component. But no way was found to know the source of the
trigger of the optionSelected event (click or keypress).

Having a flag specifiying that the autocomplete event was already
triggered thus we must ignore the keypress event will not work neither
because the flag must not be set for click events and we don't have a
way to know that the event was triggered by a click. And flag are ugly.

The only way that works was to empty the input field when an element was
added to the thesaurus list. I tried the Angular way with:
- [(ngModel)]: not working, the view's value never updated (cause was
  either of the autocomplete or the formControl attribute?).
- formControl.setValue: not working neither (cause: must be called from
  a formGroup or wathever?)

So I went the old way and got a direct reference on the component.