Restore hover effect in collection list #9432
Because list items are initialized without a parent list, they assume
that they are part of a non-interactive list, and thus never enable the
hover effect. We work around that by adding the class ourselves.

We still don't get the ripple effect, but we never had it, even before
MDC. So this is not a regression.