Hide delete button when accessing thesaurus from card
Some users accidentally press the delete button from a thesaurus details
dialog accessed from a card thinking it will remove the thesaurus from
the card.

This button remove the thesaurus from database and are not reversible,
causing some trouble to restore the deleted thresaurus.

We don't show the delete button if the detail dialog of a thesaurus is
open from a card.