Adapt thesaurus chip font size
With Angular Material 15, typography "levels" body-1 and body-2
swapped their style.

body-2 was for bolder text and body-1 was for... body.

But now body-2 are the default and body-1 for bigger text (h4).

Using default font-size for chip (14px) because text are too big
otherwise.

Source: https://material.angular.io/guide/typography#using-typography-styles-in-your-application