Always trigger Doctrine lazy loading mechanism #9494
PHP allows accessing private member of another instance from the same class. But
if we do that, it is possible that we are accessing an instance of lazy-loaded
object. And because Doctrine Proxy mechanism only triggers when accessing public
methods or public properties (but never anything private), then it means we
might access a non-loaded object. That obviously leads to all kinds of issues.

To be on the safe side, the rule of thumb is to **always use public methods** or
public properties to access things on a Doctrine entity, even if the language
could allow a shortcut.