index.html is cached for 6 hours #9796
If we don't specify anything, then browsers will make up some value. And chances
are that when we do change index.html, then browsers will not "see" it for a
while, possibly leading to mismatching JS and CSS or more severe issues.

From now on, index.html is cached for a maximum of 6 hours before re-validation.
That means if we release by night, browsers are naturally up-to-date in the
morning. And they use the cached version for most of the day.