PHP 8.2 (#138)
* PHP 8.2 #9870

* ACL SQL optimization #9743

Returning an empty string is the proper way to mean "can access to
everything" and it is more optimized than the alternative
`$this->getAllIdsQuery()`, because it does not force MariaDB to do one
more sub-query.

So the final SQL query is changed like:

```diff
- SELECT * FROM post WHERE blog_id IN (SELECT id FROM blog)
+ SELECT * FROM post
```

* Inherit Felix dependencies to ease upgrade of Felix

* Upgrade natural

* Not found entity will throw a `UserError`, not an `Error`

This is because the error category "graphql" is reserved for errors
produced by query parsing or validation. We should never have used it.

So now the error will be of category "user". And it becomes easier to
difference between broken GraphQL syntax and merely invalid values.

* Expect the correct error category

* Resilient to URL forging #9890

Try to not send total rubbish to server by doing _some_ validation of
URL/session storage data. It's not bulletproof, but it should help a
bit, by ignoring clearly invalid values.

* Install mariadb client in GitHub Actions

Because Felix requires mariadb commands