Revert "Run composer update"
This reverts commit a456791ef45ea9e01aaf51d9706af30034c2d635.

The world is not ready for this.

webonyx/graphql-php v15.8.0 don't pass PHP stan validation.
I suspect this is because they updated PHP stan dependency and we have
some problem with typing already existing but not discovered by PHP
Stan till now:

Offset 'query' ((callable():
            (GraphQL\Type\Definition\ObjectType|null))|GraphQL\Type\Definition\ObjectType|null) does not accept
            type Closure(): GraphQL\Type\Definition\NamedType.

NamedType is implemented by ObjectType and not the opposite.
NamedType cannot be casted into ObjectType because ObjectType is the
specialised class here.

Don't want to lose more time on this as PHP8.3 is anyway not compatible
with this project yet.