Angular 17 and all JS deps (#150)
* Upgrade to latest prettier #10134

* Angular 17 and all JS deps #10134

Except:

- @types/node, we run on node 18
- ngx-quill, because v25 is for quill v2 which is still RC
- ngx-scrollbar, because v14 is in beta

* Migrate to application builder #10134

* Migrate to new control flow syntax #10134

* Update eslint rules #10134

It's been a while since we updated our rules. This time we opt in into
all rules and explicitly disable the ones that don't fit our code. This
way we will automatically receive new rules in the future, and we can
decide whether we accept or reject them, instead of needing to re-check
which one is new at every update.

* Update our deps for Angular 17 #10134

* Remove empty line in imports #10134

* Re-introduce empty line to distinguish columns #10134

* Fix build output

* Update to SSR compatible gallery

* Flush pending mutation before resolving an object #10249

If we update an object, the mutation will be debounced for 2s. If we
quickly navigate to the list and click on the same object, then the
object will be re-fetched from server before the mutation is executed.
And we ultimately show a form that does not contain our latest changes.
Confusing the user that may then further overwrite other changes.

And always resolve object without cache at all

* Can exec `esbuild` after `eco permissions`

* Use singular `styleUrl`

---------

Co-authored-by: David M <david.miserez@unil.ch>