[hotfix] Optimize related cards relation query
The query generated for retrieving related cards was very poorly
optimized. It didn't take advantages of indexes by the way it
was generated and it takes ages to load related cards where it
should not be.

This commit change the filter attribute given to the backend to generate
a query using `join` instead of `where exists` and thus uses database
indexes.

The generated query will looks like:

SELECT ...
FROM
	card c0_
	INNER JOIN card_card c2_ ON c0_.id = c2_.card_source
	INNER JOIN card c1_ ON c1_.id = c2_.card_target
WHERE
	c1_.id = '6004';

Instead of:

SELECT ...
WHERE
EXISTS (
	SELECT 1
	FROM card_card c3_
	WHERE
		c3_.card_source = c0_.id
		AND c3_.card_target IN('6004')
)