Resolve card inside component instead of route
Until now, cards were loaded during navigation and not within the
component. This means that the component is only displayed once the card
is loaded. The issue is that displaying the gallery, especially when
search terms are specified, becomes increasingly slow due to the large
amount of data. So much so that when clicking the button to display the
details of an image, it sometimes takes several seconds to respond. The
cause is the number of ongoing requests that block the loading of the
card data and thus the display of the component. The user receives no
feedback that their click has been registered. We're fixing this by
loading the card data within the component and directly displaying the
card details with a loading screen.