Fix JS errors when editing objects
When migrating to Angular 17, we "fixed" templates that Angular reported
as doing useless type checking. So we typically patched with
something like:

```dif
- data.item?.permissions?.update
+ data.item.permissions.update
```

But it turns out our typing declaration was incorrect, and Angular
conclusions were incorrect. So our Angular 17 update broke the
templating in a few places.

This new patch fixes the typing, and thus re-introduces the type-checking
in the template, but in a different way.

`AbstractDetailDirective.data` was declared as being only coming from
server. Now it is declared as either coming from server (for update), or
coming from `modelService.getDefaultForServer()` (for creation). And thus,
we need to be able to narrow the type to be able to access properties
that would only be available when fetched from DB. This is what the brand
new `AbstractDetailDirective.isUpdatePage()` is used for.

So our new patch is something like:

```dif
- data.item.permissions.update
+ isUpdatePage() && data.item.permissions.update
```