[Hotfix] Fix `ghost` selection in list view
When selecting cards, doing mass edit will unselect all cards. But if
you select another card, old selected cards are still selected in the
SelectionModel but won't show the selection visually. So doing another
mass edit action will result on changing old selected cards too.

This bug occurs because there are two arrays of selected cards, a
SelectionModel in the list view and a selected array in the parent view.
When updating the selected array in the parent, the child will not
update its own SelectionModel, resulting in inconsistent behaviour.

This is not the best implemented fix here. This SelectionModel seems a
bit overkill and should be replaced with a simpler mechanism in the
futur.  Using ngOnChanges to track the parent changes on the selected
array will be far better too.

The fix here make the assumption that every action on the parent that
update selected array will change the cards array too. So we update the
SelectionModel in the child each time the dataSource is triggered. But
it is not a long term solution because if something changes only the
selected array from the parent (without triggering a dataSource change),
the child will not be updated.