Natural 57 (#155)
* Refactor resolvers to watch model from Apollo #9844

* NaturalAbstractModelService does not require passing its dependencies #9844

* Natural 57 #9844

* Return newly created item, instead of old input #9844

Should not have any differences whatsoever, but it makes more sense that
 way

* ModelService never merge mutation results into their input #9844

* Fixed typing #9844

* Refresh model name when mutating #9844

In the most common use-case the page header and the page title contain
the model `name` value coming straight from Apollo cache. That means we
need to keep that cache up to date by fetching at the very least `id`
and `name` when mutating our model.

So a typical update mutation would look like:

```graphql
updateFoo(id: $id, input: $input) {
    id
    name
    updateDate
    updater {
        ...UserMeta
    }
}
```
Go to home after tab sync #9844
Refresh model name when mutating #9844

In the most common use-case the page header and the page title contain
the model `name` value coming straight from Apollo cache. That means we
need to keep that cache up to date by fetching at the very least `id`
and `name` when mutating our model.

So a typical update mutation would look like:

```graphql
updateFoo(id: $id, input: $input) {
    id
    name
    updateDate
    updater {
        ...UserMeta
    }
}
```

This also implies that AbstractDetail never merges mutation results into
form anymore, and instead we must re-implement that mechanism on a
as needed basis. But turns out it is extremely rare that we need to do
that.

* Don't merge things for no purpose #9844

* Don't re-specify default values of mapping #10094

* `extensions` might not exist in errors #10310

The typing says `extensions` must always exist. However, it is incorrect,
because in real life runtime it is sometimes missing. This might be
something wrong in `@apollo/client`, but I could not find the root cause.

* Rename file for coherence across projects and main purpose #10310

* Avoid CLS issue on mobile #10305

On mobile the menu is closed by default. Before, it was opened, then
very quickly closed. That leads to a huge Cumulative Layout Shift issue
that is reported by Lighthouse.

---------

Co-authored-by: David M <davidmiserez@gmail.com>