Show table headers even if no result #10361
There are two reasons for this change. First, and most important, is
that a `mat-table` must never be removed from the DOM or else it will
stop listening to data. That means that if a human searches for
something but finds no result, then they will never be able to see
results anymore, unless if they reload the page.

Second, from an UX perspective, it _might_ be preferable to have a less
surprising by always showing the headers, that never change. And the
only "moving" part is the data itself, not the surrounding structure.