Each API field definition is lazy loaded #10525
Previously, we lazy-loaded the whole group of fields of QueryType and
MutationType at once. But each field were not lazy-loaded individually.
So a query with a single field, still loaded the schema for all possible
queries. Same for mutations.

Now we enforce lazy-loaded individual fields via PhpStan extra typing.
The measured speed improvement is actually extremely small, and possibly
even non-existent (!), even for our largest projects. Something like
1.01 ± 0.09 times faster than before.

But we still keep it, because the work is already done, and it
gives other minor advantages, such as:

- less indented code
- easier access to specific mutation via name instead of numerical index
- _maybe_ less memory usage (though I didn't measure it)