Try to use fulltext search when possible
This commit add fulltext indexes on multiple fields to improve search
performance.

The search will check from the Doctrine metadata the existence of a
fulltext index for the corresponding fields and use it accordingly.

It will fallback to the classic (LIKE) search if not.

On top of that, id fields will now be searched with indexes instead
of LIKE.

This fulltext search respond to the fact that on "larges" (> 200 000
cards) data, it is currently nearly unusable has it take > 15 seconds
to make a search.

A new migration is created to add the indexes. It is strongly
recommended to optimize the affected tables.

```sql
OPTIMIZE TABLE card;
OPTIMIZE TABLE country;
OPTIMIZE TABLE domain;
OPTIMIZE TABLE institution;
```