Fix multiple search terms in the same field
With the new search system, when we search multiple terms in the same
field (custom field), it will try to find all the terms in a single
attribute.

It didn't worked this way. Every terms need to be found in at least one
attribute, but terms were not grouped together.

In other words, when we search multiple terms in the same field, it must
but treated as if we search one word in multiple field.

Assum we have

Card = {name: "Chateau des Papes", locality: "Avignon"}

Before the fix:
'"Papes", "Avignon"' will return the card but "Papes Avignon" will not.

After the fix both searches will return the card.