Remove full text special chars from searches
Some characters like *, + or - have special meaning in full text search.
We do not want these functionality for the user.

If we type "Notre-Dame" we do not want the word "Notre" without the word
"Dame". So we remove all these characters.

On top of that, if we only search "Notre-" (it occurs when doing
Institution searches because it is searching while typing), a server
error occurs because full text search do not accept "-" at the end of a
word.