Fix plain_name field not filled in correctly
Since ed5df283ec3ce40c806d6b66753d19a49b509d13, a bug was introduced
that will prevent the field card.plain_name to be filled in when a card
name was specified.

This commit fix the bug and add a one time use script that will correct
the data in database.

Run it with `php bin/oneshot/fix-plain-name.php`.

The impact of this bug is minor as this field is only used for labels
in related card and quizz, and is used for sorting cards by name.