Chunk splittable facets #10557
Facets are now split into several independent lists, so each list of
facets may be split into a different chunk during production builds.
Thus reducing the Initial Total metric.

This also is strongly typed, so non-existing facets lead to compile
time error. Thanks to this, we found cases where facets incorrectly
disappeared during a refactoring of facet names.

That means that facet lists must be called in the component constructor
(injection context). In the rare cases where a facet list is used later
than the constructor, then it should be declared as a field and
initialized in the constructor.