Remove placeholder attribute from image related cards
Most of the time, it leads to performance issues rather than improving
UX.

When the system has to generate the thumbnail for the placeholder
(which is a 30px version of the image), it can take much longer
than just downloading the original picture.

We could generate the 30px version at card creation but we chose to just
remove this functionality has it is not justifying creating another
thumbnail for its purpose.