Change images height for list view
We set the same size than for the grid view so the thumbnails won't need
to be generated on the fly between the two views.

This has multiple advantages:
 - Less server overload when switching to the list view that need to
   generated on the fly a lot of thumbnails at once. Sometimes they are
   very large tiff images that take time to process and retrieve from
   the mount.
 - The different caches (browser, mount, etc) will be more efficient as it
   will need to take care of less images (if a user switch between the
   two view, all images will be retrieved from the browser cache because
   they will be the same).

Disadvantage are that it will be less efficient for list view as it need to
download larger images. But for a set of 25 images (the default item by
pages), we talk about 100-200kB.

This change is necessary because some time the server will "freeze" for
minutes, restricting users to load any page, because someone opened the
list view. It is mostly because of the time needed to retrieve all images
from the mount. On Tiresias they are a lot of tiff images > 30MB. So the
server has to retrieve mostly 1GB (25*30) data and generate thumbnails with
it.

Another approach would be to generate the thumbails after card creation.
The problem is that old cards would not have these thumbnails created and
making them with a script would be a way too big operation.