Disable PHP session locking
All images and thumbnails are stored on a NAS accessed by a mount.

Each request to access an image go through PHP. When scrolling through
the picture gallery, we do 10-20 requests simultaneously. Can be up to
50 in the list view.

These requests cannot execute themself in parallel due to PHP session
locking. Retrieving image from the NAS take time, not much, but it can
be around 100ms to 1s. Retrieving 20 images sequentially could lead to
very long loading time.

So we disable PHP session locking to allow these images to be retrieved
in parallel.

Unblocking PHP session can lead to inconsistancies, but till now they
are only used to store the auth user id and should not interfer with the
app.

!! You need to manually update local.php config file on server to use
this new configuration.