Thesaurus matchfromstart search max length limit
The "matchFromStartAutocomplete" parameter causes issues for multi-part
names. For example, with "John Doe" it's no longer possible to find the
artist by typing "doe" which is not intuitive. To avoid this problem,
we limit this feature to a maximum number of characters. Once that limit
is reached, the search behavior returns to its previous form.