Angular 19 (#185)
* Angular 19.2 and all JS deps #10723

* Enable strictStandalone #10723

* Use TypeScript module preserve option #10723

https://github.com/angular/angular-cli/commit/03180fe0358662f8fd3255ad546994da3e3bda9c

* Migrate to signal queries #10723

Because they are way easier to use.

`ng generate @angular/core:signal-queries-migration`

* Migrate to output function #10723

Because they are easier to use.

`ng generate @angular/core:output-migration`

* Upgrade natural for Angular 19.2 compatbility #10723

And refresh entire lock file

* Drop unused viewChild #10723

* viewChild might actually not exist because of `@if` #10723

* Enable `prefer-contextual-for-variables` #10723

* Partially enable `prefer-signals` #10723

Automated migration to signal input is bugged and breaks too much code
to be fixed manually. So we cannot auto-migrate. But we can still
partially enable it for `preferQuerySignals` and
`preferReadonlySignalProperties`

* Enable `no-deprecated` #10723

* Use /mnt/data/tmp for temporary build files if it exists #11323

* Restore mat-elevation-z* and hover effect on selectbox #11331

Should not have trusted automated migration during Angular 19.2
migration...

* Scrutinizer run on node 22

* Upgrade to node 22 in docker dev env

* Add empty prerendered-routes.json file

Angular 19 create this file even with not used.

---------

Co-authored-by: David M. <david.miserez@unil.ch>
Co-authored-by: David M <davidmiserez@gmail.com>
Co-authored-by: root <root@localhost>