Doctrine ORM 3 (#187)
* Mostly native DBAL enums #10570

* All unit tests pass #10570

* Happier Stan #10570

* Migrate to monolog and symfony/mailer #8635 #10257 #10724

Because laminas-log and laminas-mail are both discontinued. And we do it
now, because the migration to doctrine 3 exposed a cyclic dependency
issue in Epicerio. The end result was we were trying to log
something to DB before a connection to DB was established, and so we
established a second connection to DB, wrecking havoc in our apps.

So instead of deep diving in laminas-log to sort out all the details,
I'd rather deep dive in monolog directly and clear the technical debt at
the same time. And since monolog has an integration with symfony/mailer,
it seemed natural to do that too, and clear even more technical debt.

And to avoid the double-connection issue, the DB logger is only active,
and only retrieves its dependencies, after we are certain that a
connection to DB was established.

The overall concept of our logs remains unchanged though. We have 3 log
destinations: File, Mail and DB. Mail and DB have extra information
through `RecordCompleter`. File does not.

* Can log things to DB, even if not logging SQL queries to file #8635

The middleware is unconditionally enabled, so it can notify the
DbHandler whenever we get an active connection. However, the middleware
may or may not log SQL queries to files. This is controlled via the new
config key `logSql`.

* Load Google Map only when needed #10570

This might reduce Google Map cost, but it at least save users bandwidth
in many cases (when the card is readonly and has no geolocation)

* Leverage SMTP syntax from Symfony Mailer

* Upgrade @ecodev/natural

Without any particular reason

* Prevent `doctrine_migration_versions` to be shown as non-sync error #11390

`./bin/doctrine orm:validate-schema` and `./bin/doctrine
orm:schema-tool:update --dump-sql` both incorrectly showed the DB
structure was out of sync because of the existence of the table
`doctrine_migration_versions`. The table must exist, and it must never
be deleted.

We reproduce the behavior of Doctrine older versions to ignore that
specific table for those two specific commands.

This also happens to create a base for a future Felix-dedicated CLI with
some commands defined within Felix and others within the application.

* Don't reuse stale cache for ACL filer #11390

FilterCollection must be aware that our ACL filter internal state has
change whenever we change user. So that the Doctrine is able to properly
handle its own cache and not re-use an incorrect cache entry for the
wrong user.

This should not be a huge deal in real, production, use-cases. But it
very easily breaks unit tests where we change user often within the same
process.

* Use new signal API of Natural

* Update Angular for fixed input signal migration #11434

* Can use ST_DISTANCE with MariaDB

* Restore allowedHosts for dilps.lan and tiresias.lan

Otherwise, vite prevents us from accessing http://tiresias.lan:4400/

* Update Google IP

---------

Co-authored-by: David M. <david.miserez@unil.ch>